# backend/src/swagger.json
{
   "openapi": "3.0.0",
   "info": {
      "title": "Space App",
      "description": "Final fullstack project for Okten School",
      "version": "1.0.0"
   },
   "servers": [
      {
         "url": "http://localhost:5400/api",
         "description": "Express"
      }
   ],
   "paths": {
...
      "/tasks": {
         "post": {
            "summary": "get all notes",
            "tags": [
               "Tasks"
            ],
            "operationId": "getAllTasks",
            "security": [
               {
                  "Authorization": []
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "planId": {
                              "type": "string",
                              "example": "63ffca12826b87decc6fee24"
                           }
                        }
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Tasks"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Default error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/tasks/add": {
         "post": {
            "summary": "add task",
            "tags": [
               "Tasks"
            ],
            "operationId": "addTask",
            "security": [
               {
                  "Authorization": []
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "title": {
                              "type": "string",
                              "example": "Project"
                           },
                           "planId": {
                              "type": "string",
                              "example": "63ffca12826b87decc6fee24"
                           }
                        }
                     }
                  }
               }
            },
            "responses": {
               "201": {
                  "description": "New task was created",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Task"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Default error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/tasks/{taskId}": {
         "patch": {
            "summary": "update task by id",
            "tags": [
               "Tasks"
            ],
            "operationId": "updateTaskById",
            "parameters": [
               {
                  "name": "taskId",
                  "in": "path",
                  "required": true,
                  "description": "task id",
                  "schema": {
                     "type": "string"
                  },
                  "example": "63dfe16eda233c96fc6e2604"
               }
            ],
            "security": [
               {
                  "Authorization": []
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "isCompleted": {
                              "type": "boolean",
                              "example": true
                           }
                        }
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Success"
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid ObjectID",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ObjectIdError"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "404": {
                  "description": "Not found",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/NotExistError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         },
         "delete": {
            "summary": "delete task by id",
            "tags": [
               "Tasks"
            ],
            "operationId": "deleteTaskById",
            "parameters": [
               {
                  "name": "taskId",
                  "in": "path",
                  "required": true,
                  "description": "task id",
                  "schema": {
                     "type": "string"
                  },
                  "example": "63dfe16eda233c96fc6e2604"
               }
            ],
            "security": [
               {
                  "Authorization": []
               }
            ],
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Success"
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid ObjectID",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ObjectIdError"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "404": {
                  "description": "Not found",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/NotExistError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         }
      },
...
   },
   "components": {
      "securitySchemes": {
         "Authorization": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
         }
      },
      "schemas": {
         "UnauthorizedError": {
            "type": "object",
            "required": [
               "statusCode",
               "message"
            ],
            "properties": {
               "statusCode": {
                  "type": "number",
                  "example": 401
               },
               "message": {
                  "type": "string",
                  "example": "Invalid or expired token"
               }
            }
         },
         "ObjectIdError": {
            "type": "object",
            "required": [
               "statusCode",
               "message"
            ],
            "properties": {
               "statusCode": {
                  "type": "number",
                  "example": 400
               },
               "message": {
                  "type": "string",
                  "example": "Object ID is not valid"
               }
            }
         },
         "NotExistError": {
            "type": "object",
            "required": [
               "statusCode",
               "message"
            ],
            "properties": {
               "statusCode": {
                  "type": "number",
                  "example": 404
               },
               "message": {
                  "type": "string",
                  "example": "Object does not exist"
               }
            }
         },
         "DefaultError": {
            "type": "object",
            "properties": {
               "statusCode": {
                  "type": "number",
                  "example": 500
               },
               "message": {
                  "type": "string",
                  "example": "Error message"
               }
            }
         },
         "Success": {
            "type": "object",
            "properties": {
               "message": {
                  "type": "string",
                  "example": "Success"
               }
            }
         },
...
         "Task": {
            "type": "object",
            "properties": {
               "id": {
                  "type": "string",
                  "example": "63dfe16eda233c96fc6e2604"
               },
               "title": {
                  "type": "string",
                  "example": "Learn English"
               },
               "isCompleted": {
                  "type": "boolean",
                  "example": false
               }
            }
         },
         "Tasks": {
            "type": "array",
            "items": {
               "$ref": "#/components/schemas/Task"
            }
         },
...
      }
   }
}